Remove shallow SpockkIrFactory in favor of the more generic ContextAwareIrFactory
