+++ spockk-compiler-plugin/src/main/kotlin/io/github/pshevche/spockk/compilation/transformer/FeatureRewriter.kt
@@ -14,12 +14,23 @@
 
 package io.github.pshevche.spockk.compilation.transformer
 
+import io.github.pshevche.spockk.compilation.common.FeatureBlockStatements
 import io.github.pshevche.spockk.compilation.common.SpockkTransformationContext.FeatureContext
+import io.github.pshevche.spockk.compilation.ir.ContextAwareIrFactory
 import io.github.pshevche.spockk.compilation.ir.mutableStatements
 import org.jetbrains.kotlin.ir.declarations.IrFunction
+import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
+import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.name.Name
 
-internal class FeatureRewriter(private val irFactory: SpockkIrFactory) {
+internal class FeatureRewriter(private val irFactory: ContextAwareIrFactory) {
+
+  companion object {
+    private const val FEATURE_METADATA_FQN = "org.spockframework.runtime.model.FeatureMetadata"
+    private const val BLOCK_METADATA_FQN = "org.spockframework.runtime.model.BlockMetadata"
+    private const val BLOCK_KIND_FQN = "org.spockframework.runtime.model.BlockKind"
+  }
+
   fun rewrite(feature: IrFunction, context: FeatureContext) {
     annotateFeature(feature, context)
     renameFeature(feature, context)
@@ -28,7 +39,7 @@ internal class FeatureRewriter(private val irFactory: SpockkIrFactory) {
 
   private fun annotateFeature(feature: IrFunction, context: FeatureContext) {
     feature.annotations +=
-      irFactory.featureMetadataAnnotation(
+      featureMetadataAnnotation(
         context.ordinal,
         context.name,
         context.line,
@@ -46,4 +57,32 @@ internal class FeatureRewriter(private val irFactory: SpockkIrFactory) {
     feature.mutableStatements()?.clear()
     feature.mutableStatements()?.addAll(context.blocks.flatMap { it.statements })
   }
+
+  private fun featureMetadataAnnotation(
+    ordinal: Int,
+    name: String,
+    line: Int,
+    parameterNames: List<String>,
+    blocks: List<FeatureBlockStatements>
+  ): IrConstructorCall =
+    irFactory.constructorCall(
+      FEATURE_METADATA_FQN,
+      irFactory.const(ordinal),
+      irFactory.const(name),
+      irFactory.const(line),
+      irFactory.stringArray(parameterNames),
+      blockMetadataArray(blocks.filter { it.label.blockKind != null })
+    )
+
+  private fun blockMetadataArray(blocks: List<FeatureBlockStatements>): IrExpression =
+    irFactory.array(
+      BLOCK_METADATA_FQN,
+      blocks.map { block ->
+        irFactory.constructorCall(
+          BLOCK_METADATA_FQN,
+          irFactory.enumValue(block.label.blockKind!!, BLOCK_KIND_FQN),
+          irFactory.stringArray(listOf(block.description))
+        )
+      }
+    )
 }
